requirements.txt
Flask==2.3.3
requests==2.26.0
python-dotenv==0.19.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Werkzeug==2.3.7
flask-cors==4.0.0
